requirements.txt
backcall==0.2.0
boto==2.49.0
boto3==1.14.18
botocore==1.17.18
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
Cython==0.29.14
decorator==4.4.2
docutils==0.15.2
gensim==3.8.3
idna==2.10
ipykernel==5.3.1
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.1
jmespath==0.10.0
joblib==0.16.0
jupyter-client==6.1.5
jupyter-core==4.6.3
kiwisolver==1.2.0
llvmlite==0.33.0
matplotlib==3.2.2
numba==0.50.1
numpy==1.19.0
pandas==1.0.5
parso==0.7.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
pywin32==228
pyzmq==19.0.1
requests==2.24.0
s3transfer==0.3.3
scikit-learn==0.23.1
scipy==1.5.1
six==1.15.0
smart-open==2.1.0
threadpoolctl==2.1.0
tornado==6.0.4
tqdm==4.47.0
traitlets==4.3.3
umap-learn==0.3.10
urllib3==1.25.9
wcwidth==0.2.5
xlrd==1.2.0
